Share the envelope chain between the RTL8720C and BK72xx implementations
