+++ esphome/components/remote_transmitter/remote_transmitter.h
@@ -73,6 +73,8 @@ class RemoteTransmitterComponent final : public remote_base::RemoteTransmitterBa
   void loop() override;
   // called from the envelope timer ISR trampoline; not part of the public API
   void advance_envelope_isr();
+  // same, for trampolines whose SDK callback carries no user argument
+  static void advance_active_isr();
 #endif
 
   Trigger<> *get_transmit_trigger() { return &this->transmit_trigger_; }
@@ -100,12 +102,20 @@ class RemoteTransmitterComponent final : public remote_base::RemoteTransmitterBa
   void *pwm_{nullptr};  // pwmout_t*, opaque here to keep the SDK header out of this shared header
 #endif
 #if defined(USE_LIBRETINY_VARIANT_RTL8720C) || defined(REMOTE_TRANSMITTER_BK_PWM)
+  // Envelope chain, shared by every family that paces transmission from a hardware timer
+  // (remote_transmitter_libretiny_isr.cpp)
   void start_isr_item_(size_t index);
   void arm_envelope_timer_(uint32_t duration_us);
   void abort_stalled_chain_();
   void deliver_completion_();
   void wait_until_idle_();
   void arm_chain_(uint32_t send_times, uint32_t send_wait);
+  // Hooks implemented per family: everything the chain needs from the hardware
+  bool envelope_ready_() const;                       // PWM claimed successfully in setup()
+  void prepare_carrier_(uint32_t carrier_frequency);  // retune period, stage mark/space levels
+  void write_envelope_level_(bool mark);              // drive carrier (mark) or idle (space)
+  void arm_one_shot_(uint32_t duration_us);           // fire advance_envelope_isr after duration_us
+  void stop_envelope_timer_();
   std::vector<int32_t> isr_data_;  // owned copy of the frame; temp_ may be re-encoded mid-flight
   volatile size_t isr_index_{0};
   volatile uint32_t isr_repeats_left_{0};
@@ -118,7 +128,6 @@ class RemoteTransmitterComponent final : public remote_base::RemoteTransmitterBa
   bool stall_aborted_{false};  // this transmission ended via abort; blocks warning clear
 #endif
 #ifdef USE_LIBRETINY_VARIANT_RTL8720C
-  void update_carrier_(uint32_t carrier_frequency);
   float isr_mark_duty_{0.0f};
   float isr_space_duty_{0.0f};
 #endif
