+++ esphome/components/remote_transmitter/remote_transmitter_rtl87xx.cpp
@@ -24,20 +24,13 @@ static const char *const TAG = "remote_transmitter";
 
 #ifdef USE_LIBRETINY_VARIANT_RTL8720C
 static constexpr uint32_t ENVELOPE_TIMER_ID = TIMER6;  // GTimer7
-// Margin past a transmission's expected duration before the chain is declared stalled
-static constexpr uint32_t STALL_MARGIN_MS = 1000;
-// Longest single one-shot armed; longer durations are chained (ROM us->tick headroom unverified)
-static constexpr uint32_t MAX_ONE_SHOT_US = 50000;
 
-// Shared envelope timer: a second gtimer_init on the same id fails silently, so all
-// instances serialize on s_active_transmitter
+// One envelope timer for all instances: a second gtimer_init on the same id fails silently,
+// so the chain serializes them (remote_transmitter_libretiny_isr.cpp)
 // NOLINTBEGIN(cppcoreguidelines-avoid-non-const-global-variables)
 static uint8_t s_pwm_tick_sources[] = {GTimer1, GTimer2, GTimer3, GTimer4, GTimer5, GTimer6, 0xff};
 static gtimer_t s_envelope_timer;
 static bool s_envelope_timer_ready = false;
-static RemoteTransmitterComponent *volatile s_active_transmitter = nullptr;
-// Deadline for the in-flight transmission (millis-based); only touched from the main task
-static uint32_t s_expected_end_ms = 0;
 // NOLINTEND(cppcoreguidelines-avoid-non-const-global-variables)
 
 static void IRAM_ATTR envelope_timer_isr(uint32_t arg) {
@@ -104,105 +97,22 @@ void RemoteTransmitterComponent::digital_write(bool value) {
 }
 
 #ifdef USE_LIBRETINY_VARIANT_RTL8720C
-// Arms the shared envelope timer, chaining durations longer than MAX_ONE_SHOT_US. ISR-safe.
-void IRAM_ATTR RemoteTransmitterComponent::arm_envelope_timer_(uint32_t duration_us) {
-  // clamp to 1us (a zero-length one-shot never fires); the remainder must not underflow
-  const uint32_t chunk = std::max(uint32_t(1), std::min(duration_us, MAX_ONE_SHOT_US));
-  this->isr_wait_remaining_ = duration_us > chunk ? duration_us - chunk : 0;
-  gtimer_start_one_shout(&s_envelope_timer, chunk, (void *) envelope_timer_isr, (uint32_t) this);
-}
+// --- envelope chain hooks (see remote_transmitter_libretiny_isr.cpp) ---
 
-// Aborts a chain that stopped advancing: stop the timer, idle the pin, release the token.
-// Every step is a no-op if the chain completed meanwhile. Task context only.
-void RemoteTransmitterComponent::abort_stalled_chain_() {
-  // cleared first so a straggler one-shot bails at the ISR entry check
-  this->transmitting_ = false;
-  gtimer_stop(&s_envelope_timer);
-  pwmout_write(static_cast<pwmout_t *>(this->pwm_), this->isr_space_duty_);
-  s_active_transmitter = nullptr;
-  this->stall_aborted_ = true;
-  this->status_set_warning("envelope timer stalled");
-  ESP_LOGE(TAG, "Envelope timer stalled; transmission aborted");
-  delay(1);  // let any already-latched interrupt land while the chain state is safe
-}
+bool RemoteTransmitterComponent::envelope_ready_() const { return this->pwm_ != nullptr; }
 
-// Delivers one deferred completion with its status bookkeeping
-void RemoteTransmitterComponent::deliver_completion_() {
-  if (!this->stall_aborted_)
-    this->status_clear_warning();
-  this->complete_pending_ = false;
-  this->complete_trigger_.trigger();
-}
-
-// Writes the duty for one envelope item and arms the timer for its duration.
-// Runs in ISR context (and once from send_internal to kick the chain): no logging, no allocation.
-void IRAM_ATTR RemoteTransmitterComponent::start_isr_item_(size_t index) {
-  const int32_t item = this->isr_data_[index];
-  pwmout_write(static_cast<pwmout_t *>(this->pwm_), item > 0 ? this->isr_mark_duty_ : this->isr_space_duty_);
-  this->arm_envelope_timer_(uint32_t(item > 0 ? item : -item));
-}
-
-void IRAM_ATTR RemoteTransmitterComponent::advance_envelope_isr() {
-  if (!this->transmitting_)
-    return;  // chain was aborted; this is a stale one-shot that was already latched
-  if (this->isr_wait_remaining_ > 0) {
-    // continue a duration longer than one hardware one-shot
-    this->arm_envelope_timer_(this->isr_wait_remaining_);
-    return;
+// Retunes the PWM period when the carrier changes and stages the per-item duties;
+// unmodulated protocols (no carrier or 100% duty) drive the pin constantly during marks
+void RemoteTransmitterComponent::prepare_carrier_(uint32_t carrier_frequency) {
+  float mark_duty =
+      (carrier_frequency > 0 && this->carrier_duty_percent_ < 100) ? this->carrier_duty_percent_ / 100.0f : 1.0f;
+  float space_duty = 0.0f;
+  if (this->pin_->is_inverted()) {
+    mark_duty = 1.0f - mark_duty;
+    space_duty = 1.0f;
   }
-  if (this->isr_in_gap_) {
-    // inter-repeat gap elapsed; restart the item chain
-    this->isr_in_gap_ = false;
-    this->isr_index_ = 0;
-    this->start_isr_item_(0);
-    return;
-  }
-  this->isr_index_++;
-  if (this->isr_index_ < this->isr_data_.size()) {
-    this->start_isr_item_(this->isr_index_);
-    return;
-  }
-  // end of one repetition
-  pwmout_write(static_cast<pwmout_t *>(this->pwm_), this->isr_space_duty_);
-  if (this->isr_repeats_left_ > 1) {
-    this->isr_repeats_left_--;
-    this->isr_index_ = 0;
-    if (this->isr_send_wait_ > 0) {
-      this->isr_in_gap_ = true;
-      this->arm_envelope_timer_(this->isr_send_wait_);
-    } else {
-      this->start_isr_item_(0);
-    }
-    return;
-  }
-  this->transmitting_ = false;
-  s_active_transmitter = nullptr;
-}
-
-// Waits until no chain is in flight, delivering any deferred completions; a completion
-// automation may start a new send, so repeat until truly idle. Bounded by the stall deadline.
-void RemoteTransmitterComponent::wait_until_idle_() {
-  while (true) {
-    while (true) {
-      // snapshot: the final ISR can clear the volatile pointer between a check and a use
-      auto *active = s_active_transmitter;
-      if (active == nullptr)
-        break;
-      if ((int32_t) (millis() - s_expected_end_ms) > 0) {
-        active->abort_stalled_chain_();
-        break;
-      }
-      App.feed_wdt();
-      delay(1);
-    }
-    if (!this->complete_pending_)
-      break;
-    this->deliver_completion_();
-  }
-}
-
-// Retunes the PWM period when the carrier changes; the ISR sets duty per item
-void RemoteTransmitterComponent::update_carrier_(uint32_t carrier_frequency) {
+  this->isr_mark_duty_ = mark_duty;
+  this->isr_space_duty_ = space_duty;
   if (carrier_frequency == 0 || carrier_frequency == this->current_carrier_frequency_)
     return;
   // round(1000000/freq), clamped so a bad lambda can't hand the SDK a zero period
@@ -211,97 +121,15 @@ void RemoteTransmitterComponent::update_carrier_(uint32_t carrier_frequency) {
   this->current_carrier_frequency_ = carrier_frequency;
 }
 
-// Stages the repeat schedule and stall deadline, then starts the interrupt chain
-void RemoteTransmitterComponent::arm_chain_(uint32_t send_times, uint32_t send_wait) {
-  this->isr_repeats_left_ = send_times;
-  this->isr_send_wait_ = send_wait;
-  this->isr_index_ = 0;
-  this->isr_in_gap_ = false;
-  this->stall_aborted_ = false;
-  uint64_t frame_us = 0;
-  for (int32_t item : this->isr_data_)
-    frame_us += uint32_t(item > 0 ? item : -item);
-  const uint64_t total_us = frame_us * send_times + uint64_t(send_wait) * (send_times - 1);
-  s_expected_end_ms = millis() + uint32_t(total_us / 1000) + STALL_MARGIN_MS;
-  this->transmitting_ = true;
-  s_active_transmitter = this;
-  this->start_isr_item_(0);
+void IRAM_ATTR RemoteTransmitterComponent::write_envelope_level_(bool mark) {
+  pwmout_write(static_cast<pwmout_t *>(this->pwm_), mark ? this->isr_mark_duty_ : this->isr_space_duty_);
 }
 
-void RemoteTransmitterComponent::send_internal(uint32_t send_times, uint32_t send_wait) {
-  if (this->pwm_ == nullptr) {
-    ESP_LOGW(TAG, "Cannot send: PWM not initialized");
-    return;
-  }
-  this->wait_until_idle_();
-  if (send_times == 0) {
-    // parity with the loop-based implementations: transmit nothing, but both triggers
-    // still fire so an on_complete-sequenced automation does not stall
-    this->transmit_trigger_.trigger();
-    this->deliver_completion_();
-    return;
-  }
-  ESP_LOGD(TAG, "Sending remote code");
-  const uint32_t carrier_frequency = this->temp_.get_carrier_frequency();
-  // unmodulated protocols (no carrier or 100% duty) drive the pin constantly during marks
-  float mark_duty =
-      (carrier_frequency > 0 && this->carrier_duty_percent_ < 100) ? this->carrier_duty_percent_ / 100.0f : 1.0f;
-  float space_duty = 0.0f;
-  if (this->pin_->is_inverted()) {
-    mark_duty = 1.0f - mark_duty;
-    space_duty = 1.0f;
-  }
-  this->update_carrier_(carrier_frequency);
-  // own copy: with non_blocking the caller may re-encode temp_ while this frame is in flight
-  this->isr_data_.assign(this->temp_.get_data().begin(), this->temp_.get_data().end());
-  if (this->isr_data_.empty()) {
-    ESP_LOGW(TAG, "Empty data");
-    this->transmit_trigger_.trigger();
-    this->deliver_completion_();
-    return;
-  }
-  this->isr_mark_duty_ = mark_duty;
-  this->isr_space_duty_ = space_duty;
-  // trigger first: the deadline computed in arm_chain_ must not be charged for user code
-  this->transmit_trigger_.trigger();
-  // the automation may have started a send on another instance; let it finish before
-  // claiming the shared timer (a same-instance send remains unsupported here)
-  this->wait_until_idle_();
-  this->arm_chain_(send_times, send_wait);
-  if (this->non_blocking_) {
-    this->complete_pending_ = true;
-    this->enable_loop();
-    return;
-  }
-  // blocking mode: wait out the chain, bounded by the stall deadline
-  while (this->transmitting_) {
-    if ((int32_t) (millis() - s_expected_end_ms) > 0) {
-      this->abort_stalled_chain_();
-      break;
-    }
-    App.feed_wdt();
-    delay(1);
-  }
-  this->deliver_completion_();
+void IRAM_ATTR RemoteTransmitterComponent::arm_one_shot_(uint32_t duration_us) {
+  gtimer_start_one_shout(&s_envelope_timer, duration_us, (void *) envelope_timer_isr, (uint32_t) this);
 }
 
-void RemoteTransmitterComponent::loop() {
-  if (!this->complete_pending_) {
-    this->disable_loop();
-    return;
-  }
-  if (this->transmitting_) {
-    // non-blocking stall recovery: without this, a dead chain would leave the carrier
-    // driven and on_complete unfired until the next send happened to abort it
-    if ((int32_t) (millis() - s_expected_end_ms) <= 0)
-      return;
-    this->abort_stalled_chain_();
-  }
-  // release the loop before user code runs: the automation may start a new non-blocking
-  // send, and its enable_loop() must be the last writer or its completion would strand
-  this->disable_loop();
-  this->deliver_completion_();
-}
+void IRAM_ATTR RemoteTransmitterComponent::stop_envelope_timer_() { gtimer_stop(&s_envelope_timer); }
 
 #else  // !USE_LIBRETINY_VARIANT_RTL8720C -- AmebaZ (RTL8710B): spin-based envelope, per-frame priority boost
